+++ hydra-cardano-api/src/Hydra/Cardano/Api.hs
@@ -65,14 +65,16 @@ import Cardano.Api as X hiding (
   TxValidityUpperBound (..),
   UTxO (..),
   Witness (..),
-  multiAssetSupportedInEra,
+  createAndValidateTransactionBody,
+  defaultTxBodyContent,
+  makeShelleyKeyWitness,
   policyId,
   queryEraHistory,
   queryProtocolParameters,
   queryStakePools,
   querySystemStart,
-  scriptDataSupportedInEra,
   scriptLanguageSupportedInEra,
+  signShelleyTransaction,
   toLedgerUTxO,
  )
 import Cardano.Api.Byron as X (
@@ -116,6 +118,7 @@ import Hydra.Cardano.Api.Prelude (
 
 import Hydra.Cardano.Api.Address ()
 import Hydra.Cardano.Api.AddressInEra as Extras
+import Hydra.Cardano.Api.AlonzoEraOnwards as Extras
 import Hydra.Cardano.Api.BlockHeader as Extras
 import Hydra.Cardano.Api.ChainPoint as Extras
 import Hydra.Cardano.Api.CtxTx as Extras
@@ -124,13 +127,12 @@ import Hydra.Cardano.Api.ExecutionUnits as Extras
 import Hydra.Cardano.Api.Hash as Extras
 import Hydra.Cardano.Api.KeyWitness as Extras
 import Hydra.Cardano.Api.Lovelace as Extras
-import Hydra.Cardano.Api.MultiAssetSupportedInEra as Extras
+import Hydra.Cardano.Api.MaryEraOnwards as Extras
 import Hydra.Cardano.Api.NetworkId ()
 import Hydra.Cardano.Api.PlutusScript as Extras
 import Hydra.Cardano.Api.PolicyId as Extras
 import Hydra.Cardano.Api.ReferenceScript as Extras
 import Hydra.Cardano.Api.ScriptData as Extras
-import Hydra.Cardano.Api.ScriptDataSupportedInEra as Extras
 import Hydra.Cardano.Api.ScriptDatum as Extras
 import Hydra.Cardano.Api.ScriptHash as Extras
 import Hydra.Cardano.Api.ScriptLanguageInEra as Extras
@@ -337,6 +339,9 @@ pattern ShelleyTxBody
     ShelleyTxBody =
       Cardano.Api.Shelley.ShelleyTxBody ShelleyBasedEraBabbage
 
+signShelleyTransaction :: TxBody -> [ShelleyWitnessSigningKey] -> Tx
+signShelleyTransaction = Cardano.Api.signShelleyTransaction shelleyBasedEra
+
 -- ** TxAuxScripts
 
 type TxAuxScripts = Cardano.Api.TxAuxScripts Era
@@ -354,7 +359,7 @@ pattern TxAuxScripts{txAuxScripts'} <-
   Cardano.Api.TxAuxScripts _ txAuxScripts'
   where
     TxAuxScripts =
-      Cardano.Api.TxAuxScripts AuxScriptsInBabbageEra
+      Cardano.Api.TxAuxScripts AllegraEraOnwardsBabbage
 
 -- ** TxBody
 
@@ -365,6 +370,12 @@ pattern TxBody :: TxBodyContent ViewTx -> TxBody
 pattern TxBody{txBodyContent} <-
   Cardano.Api.TxBody txBodyContent
 {-# COMPLETE TxBody #-}
+
+createAndValidateTransactionBody :: TxBodyContent BuildTx -> Either TxBodyError TxBody
+createAndValidateTransactionBody = Cardano.Api.createAndValidateTransactionBody cardanoEra
+
+defaultTxBodyContent :: TxBodyContent BuildTx
+defaultTxBodyContent = Cardano.Api.defaultTxBodyContent cardanoEra
 
 -- ** TxBodyContent
 
@@ -379,7 +390,8 @@ pattern TxBodyContent ::
   TxTotalCollateral Era ->
   TxReturnCollateral CtxTx Era ->
   TxFee ->
-  (TxValidityLowerBound, TxValidityUpperBound) ->
+  TxValidityLowerBound ->
+  TxValidityUpperBound ->
   TxMetadataInEra ->
   TxAuxScripts ->
   TxExtraKeyWitnesses ->
@@ -400,7 +412,8 @@ pattern TxBodyContent
   , txTotalCollateral
   , txReturnCollateral
   , txFee
-  , txValidityRange
+  , txValidityLowerBound
+  , txValidityUpperBound
   , txMetadata
   , txAuxScripts
   , txExtraKeyWits
@@ -410,8 +423,8 @@ pattern TxBodyContent
   , txUpdateProposal
   , txMintValue
   , txScriptValidity
-  , txGovernanceActions
-  , txVotes
+  , txProposalProcedures
+  , txVotingProcedures
   } <-
   Cardano.Api.TxBodyContent
     txIns
@@ -421,7 +434,8 @@ pattern TxBodyContent
     txTotalCollateral
     txReturnCollateral
     txFee
-    txValidityRange
+    txValidityLowerBound
+    txValidityUpperBound
     txMetadata
     txAuxScripts
     txExtraKeyWits
@@ -431,8 +445,8 @@ pattern TxBodyContent
     txUpdateProposal
     txMintValue
     txScriptValidity
-    txGovernanceActions
-    txVotes
+    txProposalProcedures
+    txVotingProcedures
   where
     TxBodyContent = Cardano.Api.TxBodyContent
 
@@ -456,7 +470,7 @@ pattern TxBodyScriptData{txBodyScriptDatums, txBodyScriptRedeemers} <-
   Cardano.Api.TxBodyScriptData _ txBodyScriptDatums txBodyScriptRedeemers
   where
     TxBodyScriptData =
-      Cardano.Api.TxBodyScriptData ScriptDataInBabbageEra
+      Cardano.Api.TxBodyScriptData AlonzoEraOnwardsBabbage
 
 -- ** TxExtraKeyWitnesses
 
@@ -474,7 +488,7 @@ pattern TxExtraKeyWitnesses{txExtraKeyWitnesses} <-
   Cardano.Api.TxExtraKeyWitnesses _ txExtraKeyWitnesses
   where
     TxExtraKeyWitnesses =
-      Cardano.Api.TxExtraKeyWitnesses ExtraKeyWitnessesInBabbageEra
+      Cardano.Api.TxExtraKeyWitnesses AlonzoEraOnwardsBabbage
 
 -- ** TxFee
 
@@ -486,7 +500,7 @@ pattern TxFeeExplicit{txFeeExplicit} <-
   Cardano.Api.TxFeeExplicit _ txFeeExplicit
   where
     TxFeeExplicit =
-      Cardano.Api.TxFeeExplicit TxFeesExplicitInBabbageEra
+      Cardano.Api.TxFeeExplicit ShelleyBasedEraBabbage
 
 -- ** TxIns
 
@@ -509,7 +523,7 @@ pattern TxInsReference{txInsReference'} <-
   Cardano.Api.TxInsReference _ txInsReference'
   where
     TxInsReference =
-      Cardano.Api.TxInsReference Cardano.Api.Shelley.ReferenceTxInsScriptsInlineDatumsInBabbageEra
+      Cardano.Api.TxInsReference Cardano.Api.Shelley.BabbageEraOnwardsBabbage
 
 -- ** TxInsCollateral
 
@@ -528,7 +542,7 @@ pattern TxInsCollateral{txInsCollateral'} <-
   Cardano.Api.TxInsCollateral _ txInsCollateral'
   where
     TxInsCollateral =
-      Cardano.Api.TxInsCollateral CollateralInBabbageEra
+      Cardano.Api.TxInsCollateral AlonzoEraOnwardsBabbage
 
 -- ** TxMetadataInEra
 
@@ -547,7 +561,7 @@ pattern TxMetadataInEra{txMetadataInEra} <-
   Cardano.Api.TxMetadataInEra _ txMetadataInEra
   where
     TxMetadataInEra =
-      Cardano.Api.TxMetadataInEra TxMetadataInBabbageEra
+      Cardano.Api.TxMetadataInEra ShelleyBasedEraBabbage
 
 -- ** TxMintValue
 
@@ -569,7 +583,7 @@ pattern TxMintValue{txMintValueInEra, txMintValueScriptWitnesses} <-
   Cardano.Api.TxMintValue _ txMintValueInEra txMintValueScriptWitnesses
   where
     TxMintValue =
-      Cardano.Api.TxMintValue MultiAssetInBabbageEra
+      Cardano.Api.TxMintValue MaryEraOnwardsBabbage
 
 -- ** TxOut
 
@@ -581,14 +595,14 @@ pattern TxOut :: AddressInEra -> Value -> TxOutDatum ctx -> ReferenceScript -> T
 pattern TxOut{txOutAddress, txOutValue, txOutDatum, txOutReferenceScript} <-
   Cardano.Api.TxOut
     txOutAddress
-    (TxOutValue MultiAssetInBabbageEra txOutValue)
+    (TxOutValue MaryEraOnwardsBabbage txOutValue)
     txOutDatum
     txOutReferenceScript
   where
     TxOut addr value datum ref =
       Cardano.Api.TxOut
         addr
-        (TxOutValue MultiAssetInBabbageEra value)
+        (TxOutValue MaryEraOnwardsBabbage value)
         datum
         ref
 
@@ -600,12 +614,12 @@ type ReferenceScript = Cardano.Api.Shelley.ReferenceScript Era
 pattern ReferenceScript :: ScriptInAnyLang -> ReferenceScript
 pattern ReferenceScript{referenceScript} <-
   Cardano.Api.Shelley.ReferenceScript
-    Cardano.Api.Shelley.ReferenceTxInsScriptsInlineDatumsInBabbageEra
+    Cardano.Api.Shelley.BabbageEraOnwardsBabbage
     referenceScript
   where
     ReferenceScript =
       Cardano.Api.Shelley.ReferenceScript
-        Cardano.Api.Shelley.ReferenceTxInsScriptsInlineDatumsInBabbageEra
+        Cardano.Api.Shelley.BabbageEraOnwardsBabbage
 
 pattern ReferenceScriptNone :: Cardano.Api.Shelley.ReferenceScript Era
 pattern ReferenceScriptNone <-
@@ -631,21 +645,21 @@ pattern TxOutDatumHash{txOutDatumHash} <-
   Cardano.Api.TxOutDatumHash _ txOutDatumHash
   where
     TxOutDatumHash =
-      Cardano.Api.TxOutDatumHash ScriptDataInBabbageEra
+      Cardano.Api.TxOutDatumHash AlonzoEraOnwardsBabbage
 
 pattern TxOutDatumInTx :: HashableScriptData -> TxOutDatum CtxTx
 pattern TxOutDatumInTx{txOutDatumScriptData} <-
   Cardano.Api.TxOutDatumInTx _ txOutDatumScriptData
   where
     TxOutDatumInTx =
-      Cardano.Api.TxOutDatumInTx ScriptDataInBabbageEra
+      Cardano.Api.TxOutDatumInTx AlonzoEraOnwardsBabbage
 
 pattern TxOutDatumInline :: HashableScriptData -> TxOutDatum ctx
 pattern TxOutDatumInline{txOutDatumInlineScriptData} <-
   Cardano.Api.TxOutDatumInline _ txOutDatumInlineScriptData
   where
     TxOutDatumInline =
-      Cardano.Api.TxOutDatumInline Cardano.Api.Shelley.ReferenceTxInsScriptsInlineDatumsInBabbageEra
+      Cardano.Api.TxOutDatumInline Cardano.Api.Shelley.BabbageEraOnwardsBabbage
 
 -- ** TxScriptValidity
 
@@ -664,7 +678,7 @@ pattern TxScriptValidity{txScriptValidity'} <-
   Cardano.Api.TxScriptValidity _ txScriptValidity'
   where
     TxScriptValidity =
-      Cardano.Api.TxScriptValidity TxScriptValiditySupportedInBabbageEra
+      Cardano.Api.TxScriptValidity AlonzoEraOnwardsBabbage
 
 -- ** TxValidityLowerBound
 
@@ -679,11 +693,11 @@ pattern TxValidityNoLowerBound <-
       Cardano.Api.TxValidityNoLowerBound
 
 pattern TxValidityLowerBound :: SlotNo -> TxValidityLowerBound
-pattern TxValidityLowerBound{txValidityLowerBound} <-
-  Cardano.Api.TxValidityLowerBound _ txValidityLowerBound
+pattern TxValidityLowerBound{lowerBound} <-
+  Cardano.Api.TxValidityLowerBound _ lowerBound
   where
     TxValidityLowerBound =
-      Cardano.Api.TxValidityLowerBound ValidityLowerBoundInBabbageEra
+      Cardano.Api.TxValidityLowerBound AllegraEraOnwardsBabbage
 
 -- ** TxValidityUpperBound
 
@@ -691,18 +705,12 @@ type TxValidityUpperBound = Cardano.Api.TxValidityUpperBound Era
 {-# COMPLETE TxValidityNoUpperBound, TxValidityUpperBound #-}
 
 pattern TxValidityNoUpperBound :: TxValidityUpperBound
-pattern TxValidityNoUpperBound <-
-  Cardano.Api.TxValidityNoUpperBound ValidityNoUpperBoundInBabbageEra
-  where
-    TxValidityNoUpperBound =
-      Cardano.Api.TxValidityNoUpperBound ValidityNoUpperBoundInBabbageEra
+pattern TxValidityNoUpperBound =
+  Cardano.Api.TxValidityUpperBound ShelleyBasedEraBabbage Nothing
 
 pattern TxValidityUpperBound :: SlotNo -> TxValidityUpperBound
-pattern TxValidityUpperBound{txValidityUpperBound} <-
-  Cardano.Api.TxValidityUpperBound _ txValidityUpperBound
-  where
-    TxValidityUpperBound =
-      Cardano.Api.TxValidityUpperBound ValidityUpperBoundInBabbageEra
+pattern TxValidityUpperBound{upperBound} =
+  Cardano.Api.TxValidityUpperBound ShelleyBasedEraBabbage (Just upperBound)
 
 -- ** Witness
 
@@ -720,3 +728,6 @@ pattern ScriptWitness scriptWitnessInCtx scriptWitness <-
   Cardano.Api.ScriptWitness scriptWitnessInCtx scriptWitness
   where
     ScriptWitness = Cardano.Api.ScriptWitness
+
+makeShelleyKeyWitness :: TxBody -> ShelleyWitnessSigningKey -> KeyWitness
+makeShelleyKeyWitness = Cardano.Api.makeShelleyKeyWitness shelleyBasedEra