+++ hydra-cardano-api/src/Hydra/Cardano/Api/AlonzoEraOnwards.hs
@@ -1,0 +1,14 @@
+module Hydra.Cardano.Api.AlonzoEraOnwards where
+
+import Hydra.Cardano.Api.Prelude
+
+-- | Type class to produce 'AlonzoEraOnwards' witness values while staying
+-- parameterized by era.
+class IsAlonzoEraOnwards era where
+  alonzoEraOnwards :: AlonzoEraOnwards era
+
+instance IsAlonzoEraOnwards AlonzoEra where
+  alonzoEraOnwards = AlonzoEraOnwardsAlonzo
+
+instance IsAlonzoEraOnwards BabbageEra where
+  alonzoEraOnwards = AlonzoEraOnwardsBabbage