+++ hydra-cardano-api/src/Hydra/Cardano/Api/BabbageEraOnwards.hs
@@ -1,0 +1,11 @@
+module Hydra.Cardano.Api.BabbageEraOnwards where
+
+import Hydra.Cardano.Api.Prelude
+
+-- | Type class to produce 'BabbageEraOnwards' witness values while staying
+-- parameterized by era.
+class IsBabbageEraOnwards era where
+  babbageEraOnwards :: BabbageEraOnwards era
+
+instance IsBabbageEraOnwards BabbageEra where
+  babbageEraOnwards = BabbageEraOnwardsBabbage