+++ hydra-cluster/src/CardanoClient.hs
@@ -50,6 +50,7 @@ calculateMinFee networkId body Sizes{inputs, outputs, witnesses} pparams =
   let tx = makeSignedTransaction [] body
       noByronWitnesses = 0
    in estimateTransactionFee
+        shelleyBasedEra
         networkId
         (protocolParamTxFeeFixed pparams)
         (protocolParamTxFeePerByte pparams)